Expand Information
My objective for this exercise was to expand the information that was displayed. First I noticed that only the first entry was being displayed so I expanded that by changing WordResponse to WordEntry and removed Word so that each entry could hold the information that would be displayed and would more accurately reflect the format that Merriam-Webster provided. With that I added the HeadWord to each entry so that each entry would be its own section in the table view. Then I added the Pronunciation cell.